refactor firebase variables
