+++ src/config.js
@@ -1,0 +1,11 @@
+const config = {
+  apiKey: "AIzaSyCZoSGSJicaK9CuthGIv0K5n17bOr9LKmM",
+  authDomain: "the-farm-223df.firebaseapp.com",
+  databaseURL: "https://the-farm-223df.firebaseio.com",
+  projectId: "the-farm-223df",
+  storageBucket: "the-farm-223df.appspot.com",
+  messagingSenderId: "1061647793143",
+  appId: "1:1061647793143:web:00780a2ea626e3d5ca4cf2",
+  measurementId: "G-NKPNKR7JKP"
+};
+export default config;
